chore: address /simplify findings on buy flow
Consolidate the Apple Pay minimum dialog into `DialogItem+CashFlows.swift`
and drop the dedicated `DialogItem+Buy.swift`. Extract the $5 Coinbase
minimum to `OnrampCoordinator.minimumPurchaseUSD` so the guard and the
formatted dialog share one source. Tighten the `appRouterNestedSheet`
doc-comment now that the trade-off is well established.
